Refactored mopidy_tidal caches
The extension previously used three different caches:

- A simple in-memory dictionary for playlists
- A more sophisticated local `Cache` for tracks
- `LruCache` (extended by `SearchCache`) for everything else

There were also many overlaps between the tracks `Cache` and `LruCache`.

This commit rationalizes the caches and implements a single `LruCache`
for all the items, with support for filesystem persistence.

Items cached on the filesystem are indexed as it follows:

```
<cache_dir>
  -> <type>       # e.g. track or playlist
    -> <prefix>   # first two characters of the ID
      -> <id>.cache
```

Where `<id>.cache` is an object representation serialized by `pickle`.