feat: login hack.
When `login_method` is set to `HACK`, Mopidy-Tidal will return fake
objects (track/playlist/artist/etc) by intercepting all user interactions.
These fake objects point to the login url.

Additionally, we use a text-to-speech service to read out the login code,
because if you're going to have a hack, you might as well have an *impressive*
one.

This hack is rather complicated as it depends on a variety of introspection
tools.  It is thoroughly tested both on the unit and integration level.